requirements.txt
Babel==2.9.0
click==7.1.2
defusedxml==0.6.0
dominate==2.6.0
Flask==1.1.2
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
lxml==4.6.2
MarkupSafe==1.1.1
python3-saml==1.10.1
pytz==2021.1
PyYAML==5.4.1
six==1.15.0
visitor==0.1.3
Werkzeug==1.0.1
xmlsec==1.3.9
